requirements.txt
amqp==5.0.2
appdirs==1.4.3
attrs==19.3.0
autopep8==1.5.4
billiard==3.6.3.0
black==19.10b0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255208697/work
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography @ file:///C:/ci/cryptography_1607637849569/work
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
marshmallow==3.10.0
pathspec==0.7.0
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
regex==2020.2.20
six @ file:///C:/ci/six_1605187374963/work
toml==0.10.0
typed-ast==1.4.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
